extract Structure inner classes and move them to 'structure' subpackage
