+++ Plugin/src/main/java/plugin/structure/Spawner.java
@@ -1,0 +1,10 @@
+package plugin.structure;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Spawner {
+    public final List<String> mobIDs = new ArrayList<>();
+    public final List<Short> weights = new ArrayList<>();
+    public int totalweight;
+}
